requirements.txt
streamlit 
opencv-python 
groq 
python-dotenv
pandas
plotly
